refactor(invoke): split streaming runtime helpers (#840)
